LayoutTests/media/modern-media-controls/skip-back-button/skip-back-button.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>SkipBackButton</code> class.");

const skipBackButton = new SkipBackButton({ layoutTraits: LayoutTraits.macOS });

shouldBeEqualToString("skipBackButton.element.localName", "button");
shouldBeTrue('skipBackButton.element.classList.contains("icon")');
shouldBeTrue('skipBackButton.element.classList.contains("skip-back")');
shouldBeEqualToString("skipBackButton.iconName", `${Icons.SkipBack}`);

const iOSSkipBackButton = new SkipBackButton({ layoutTraits: LayoutTraits.iOS });

scheduler.frameDidFire = function()
{
    if ([skipBackButton, iOSSkipBackButton].some(button => button.width == 0))
        return;

    shouldBeTrue('skipBackButton.element.style.webkitMaskImage.includes("macOS/interval-skip-back@")');
    shouldBeTrue('iOSSkipBackButton.element.style.webkitMaskImage.includes("iOS/interval-skip-back@")');
    finishJSTest();
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
